Update integ tests; rename network in docker; update README

+++ run-test-docker-image.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/env bash
 
+# A script that builds the Volttron image, runs the container which creates an instance of Volttron ZMQ, and performs basic checks on the Volttron ZMQ instance that is running on the container.
+# The purpose of this script is to give the developer a tool to check that the image works before submitting a PR. The script is also used to conduct
+# a basic integration test in the Github Actions Workflow, integ_test.yml, as part of the CI pipeline
+# NOTE: this script does not currently test Volttron on RMQ
+
 sudo apt-get install jq -y
 start=$(date +%s)
 
@@ -13,8 +18,13 @@ exit_cleanly() {
 }
 
 
-exit_test() {
+exit_on_failed_test() {
   echo -e $1
+  exit_code=$1
+  last_command=${@:2}
+  if [ $exit_code -ne 0 ]; then
+      >&2 echo "\"${last_command}\" command failed with exit code ${exit_code}."
+  fi
   docker logs --tail 25 volttron1
   docker-compose down
   exit_cleanly
@@ -23,7 +33,7 @@ exit_test() {
 check_error_code() {
   local code=$1
   if [ "${code}" -ne 0 ]; then
-    exit_test $2
+    exit_on_failed_test $2
   fi
 }
 
@@ -31,7 +41,7 @@ check_error_code() {
 skip_build='' # skip building the image
 wait=360 # 6 minutes; wait is used for sleep while the container is setting up Volttron
 group='volttron' # group name of the image; will be used to name the image <group>/volttron
-tag='develop' # image tag; will be used to name the image <source image>:<tag>
+tag='zmq' # image tag; will be used to name the image <source image>:<tag>
 while getopts 'sw:g:t:' flag; do
   case "${flag}" in
     s) skip_build=true ;;
@@ -53,12 +63,12 @@ if [ "${skip_build}" = true ]; then
 else
   echo "Building image: ${image_name}"
   git submodule update --init --recursive
-  docker rmi "${image_name}" --force
-  docker build --no-cache -t "${image_name}" .
+  docker-compose build --no-cache --force-rm volttron1
 fi
 
 ###### Test that the image was built
-docker images --format "{{.Tag}}: {{.Repository}}" | grep 'develop: volttron/volttron'
+echo 'Checking that the image was built...'
+docker images --format "{{.Tag}}: {{.Repository}}" | grep 'zmq: volttron/volttron'
 check_error_code $? 'Failed to build image'
 
 ############ Setup and start container
@@ -67,7 +77,7 @@ echo "Will try at most ${attempts} attempts to start container..."
 while [ "${attempts}" -gt 0 ]; do
   echo "Attempt number ${attempts} to start container."
   docker-compose up --detach
-  echo "Configuring and starting Volttron platform; this will take approximately several minutes........"
+  echo "Configuring and starting Volttron platform; this will take approximately $(expr ${wait} / 60) minutes........"
   sleep ${wait}
   docker ps --filter "name=volttron1" --filter "status=running" | grep 'volttron1'
   tmp_code=$?
@@ -99,14 +109,14 @@ output=$(docker exec -u volttron volttron1 "${vctl}" list)
 check_error_code $? "Failed to get list of agents"
 count=$(echo "${output}" | grep "" -c)
 if [ "${count}" -ne 6 ]; then
-  exit_test "Total count of agents were not installed. Current count: ${count} \n Output from vctl list: \n ${output}"
+  exit_on_failed_test "Total count of agents were not installed. Current count: ${count} \n Output from vctl list: \n ${output}"
 fi
 echo "PASSED....CHECKING LIST OF AGENTS."
 
 # Test
-# Check the health/status of each agent
+# Check the health/status of agents
 declare -a agents_message
-agents_message=('listeneragent-3.3' 'platform_driveragent-4.0' 'sqlhistorianagent-3.7.0')
+agents_message=('listeneragent-3.3' 'platform_driveragent-4.0')
 echo "Testing health of agents: ${agents_message[*]} "
 for agent in "${agents_message[@]}"
 do
@@ -116,7 +126,7 @@ do
   if [ "${message}" = '"GOOD"' ]; then
     echo "Agent is healthy"
   else
-    exit_test "Failing Agent Health test because agent is unhealthy: ${output}"
+    exit_on_failed_test "Failing Agent Health test because agent is unhealthy: ${output}"
   fi
 done
 echo "PASSED....CHECKING HEALTH OF AGENTS."
@@ -134,8 +144,8 @@ echo "Testing platform configuration..."
 output=$(docker exec -u volttron volttron1 cat /home/volttron/.volttron/config)
 check_error_code $? 'Failed to get platform configuration'
 count=$(docker exec -u volttron volttron1 cat /home/volttron/.volttron/config | grep "" -c)
-if [ "${count}" -ne 8 ]; then
-  exit_test "Platform not correctly configured. Expected at least 8 lines of configuration. ${output}"
+if [ "${count}" -ne 9 ]; then
+  exit_on_failed_test "Platform not correctly configured. Expected at least 8 lines of configuration. ${output}"
 fi
 echo "PASSED....CHECKING PLATFORM CONFIG."
 
@@ -143,11 +153,11 @@ echo "PASSED....CHECKING PLATFORM CONFIG."
 # Check that PlatformWeb is working by calling the discovery endpoint; the output is a JSON consisting of several keys such
 # as "server-key", "instance_name"; here we are checking "instance_name" matches the instance name that we set in platform_config.yml
 echo "Testing discovery endpoint on web..."
-output=$(curl -s 'http://0.0.0.0:8080/discovery/')
-check_error_code $? 'Failed to get or parse http://0.0.0.0:8080/discovery'
+output=$(curl -s 'https://172.28.5.1:8443/discovery/')
+check_error_code $? 'Failed to get or parse https://172.28.5.1:8443/discovery'
 instance_name=$(echo ${output} | jq .\"instance-name\")
 if [[ "$instance_name" != '"volttron1"' ]]; then
-  exit_test "Failing Discovery Test. Instance name is not correct. instance_name: ${instance_name}. Complete output ${output}"
+  exit_on_failed_test "Failing Discovery Test. Instance name is not correct. instance_name: ${instance_name}. Complete output ${output}"
 fi
 echo "PASSED....CHECKING DISCOVERY OF WEB UI."
